bugfix: callback listener registered twice
